requirements.txt
appdirs==1.4.4
mysql-connector-python==8.0.23
protobuf==3.15.6
pyee==8.1.0
pyppeteer==0.2.5
python-dotenv==0.16.0
six==1.15.0
tqdm==4.59.0
urllib3==1.26.4
websockets==8.1
